requirements.txt
numpy==1.26.4
pandas==1.5.3
PyYAML==6.0.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
SQLAlchemy==2.0.27
tabula_py==2.9.0
